Fix last cascade blur issue
When limiting the refresh frequency for last cascade,
if MSAA was off, the last cascade was getting blurred
each frame even if not rendered.

With MSAA this was not an issue due to the explicit
StretchRect to resolve MSAA which made sure that the
last cascade was always as if just rendered.