Prepare for standalone reversed depth buffer
Removed depth inversion from NVR as a switch.
Automatically detects the DI plugin and makes
sure that shaders are compiled with respect to it.

The constant used for depth buffer combination and
some checks is still present.

Should make it so NVR is compatible with both
setups with and without depth inversion.